third_party/blink/web_tests/clipboard/async-clipboard/async-selective-format-read.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>Async clipboard selective format read</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<body>Body needed for test_driver.click()</body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="resources/user-activation.js"></script>
<script>

async function WritePlainTextAndHTMLToClipboard() {
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');
  const formatPlain = 'text/plain';
  const formatHtml = 'text/html';
  const blobPlain = new Blob(['plain text content'], {type: formatPlain});
  const blobHtml = new Blob(['<p>html content</p>'], {type: formatHtml});
  const clipboardItem = new ClipboardItem({
    [formatPlain]: blobPlain,
    [formatHtml]: blobHtml
  });
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItem]);
}

async function WriteCustomFormatToClipboard() {
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');
  const formatPlain = 'text/plain';
  const formatCustom = 'web application/x-custom';
  const blobPlain = new Blob(['plain text content'], {type: formatPlain});
  const blobCustom = new Blob(['custom format data'], {type: 'application/x-custom'});
  const clipboardItem = new ClipboardItem({
    [formatPlain]: blobPlain,
    [formatCustom]: blobCustom
  });
  await waitForUserActivation();
  await navigator.clipboard.write([clipboardItem]);
}

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({types: ['text/plain']});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_array_equals(clipboard_mime_types, ['text/plain']);
}, 'Reading only plain text');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({
    types: ['text/plain'],
    unsanitized: ['text/html']
  });
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_array_equals(clipboard_mime_types, ['text/plain']);
}, 'Reading only plain text and unsanitized option is mentioned');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({
    types: ['text/html'],
    unsanitized: ['text/html']
  });
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_array_equals(clipboard_mime_types, ['text/html']);
}, 'Reading unsanitized html');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read();
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_array_equals(clipboard_mime_types, ['text/plain', 'text/html']);
}, 'Reading all types');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({types: []});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_equals(clipboard_mime_types.length, 0);
}, 'When types is empty');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({types: ['text/text']});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_equals(clipboard_mime_types.length, 0);
}, 'Reading invalid types');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({types: null});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_equals(clipboard_mime_types.length, 0);
}, 'types: null should not return any clipboard data');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({types: undefined});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_array_equals(clipboard_mime_types, ['text/plain', 'text/html']);
}, 'types: undefined should return all clipboard data');

promise_test(async t => {
  await WriteCustomFormatToClipboard();
  const items = await navigator.clipboard.read({types: ['web application/x-custom']});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_array_equals(clipboard_mime_types, ['web application/x-custom']);
}, 'Reading only custom format');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({types: ['text/plain', 'image/png']});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_array_equals(clipboard_mime_types, ['text/plain']);
}, 'Read only available types');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: 42}));
}, 'types: number should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: "text/plain"}));
}, 'types: string should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: document.createElement("div")}));
}, 'types: DOM element should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: {}}));
}, 'types: object should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  await promise_rejects_js(t, TypeError,
    navigator.clipboard.read({types: () => {}}));
}, 'types: function should throw TypeError');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({types: [() => {}]});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_equals(clipboard_mime_types.length, 0);
}, 'types: function in array should not return any clipboard data');

promise_test(async t => {
  await WritePlainTextAndHTMLToClipboard();
  const items = await navigator.clipboard.read({types: ['text/plain', 'text/plain']});
  assert_equals(items.length, 1);
  const item = items[0];
  const clipboard_mime_types = item.types;
  assert_array_equals(clipboard_mime_types, ['text/plain']);
}, 'Reading duplicate types should return single instance');
</script>
